openagents/conclave/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "conclave"
version = "0.1.0"
description = "Peer-to-peer agent conclave protocol over Gensyn AXL."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Professor Codephreak", email = "codephreak@pythai.net"},
]
keywords = ["axl", "gensyn", "p2p", "agents", "mindx", "bonafide", "yggdrasil"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: System :: Networking",
]

dependencies = [
    "httpx>=0.27",
    "pynacl>=1.5",
    "cbor2>=5.6",
    "cryptography>=42.0",
]

[project.optional-dependencies]
chain = ["web3>=6.20"]
algorand = ["py-algorand-sdk>=2.6"]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.urls]
Homepage = "https://agenticplace.pythai.net"
Repository = "https://github.com/codephreak/conclave"

[tool.hatch.build.targets.wheel]
packages = ["conclave"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
ignore_missing_imports = true
